requirements.txt
streamlit==1.31.0
opencv-python-headless==4.9.0.80
numpy==1.24.3
Pillow==10.2.0
scikit-learn==1.4.0
scipy==1.12.0
pygame==2.5.2
